Support singular file for want updates.
This change removes the multi file want updates and replaces it
with a single file, which captures all the changes for that given want.
This should be fine since the amount of updates will be small.

To support this, also adds a cleaning task to remove all the old want
updates from the database.