Aligns digital wantlist settings in wants.
This change ensures that when items are removed from the
digital wantlist that the source chip is removed from the
underlying want. This is the first step in ensuring that wants
are tied to wantlist.

This works towards #1790.

Next step is to retire wants when they have no associated wantlists.