Working on getting order by artist working correctly.
Starting with some basic ordering tests, before we think about
how we can expand this dramatically. We may need to incorporate
some corpora to learn the correct ordering since it's not
immediately obvious that The Beatles should be ordered as "Beatles, The"
but let's figure that out once we have some basic in place.

This closes #1801.